mlapi/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Mai La <mai.la@berkeley.edu>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
transformers = "^4.30.2"
torch = [
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl", "platform" = "linux"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-win_amd64.whl", "platform" = "win32"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_10_9_x86_64.whl", "platform" = "darwin",  markers = "platform_machine == 'x86_64'" },
  {url = "http://download.pytorch.org/whl/cpu/torch-2.0.1-cp310-none-macosx_11_0_arm64.whl", "platform" = "darwin",  markers = "platform_machine == 'arm64'" }
]
scipy = "^1.9"
scikit-learn = "^1.1.2"
datasets = "^2.13.1"
accelerate ="^0.20.3"
requests = "^2.28"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
fastapi = "^0.85"
uvicorn = "^0.18"
rdkit = "^2023.3.1"
numpy = "^1.22.4"
gensim="^4.2.0"
pandas = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
isort = "^5.10.1"
black = "^22.1"
pre-commit = "^2.17.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
